+++ src/parser/inline/html.test.ts
@@ -11,41 +11,41 @@ describe('Unit: parser/html', () => {
       assert.deepStrictEqual(inspect(parser('<script>alert()<script>')), void 0);
       assert.deepStrictEqual(inspect(parser('<script>alert()</script>')), void 0);
       assert.deepStrictEqual(inspect(parser('<script src="."></script>')), void 0);
-      assert.deepStrictEqual(inspect(parser('<var onclick="alert()">')), void 0);
-      assert.deepStrictEqual(inspect(parser('<var onclick="alert()"><var>')), void 0);
-      assert.deepStrictEqual(inspect(parser('<var><var onclick="alert()"></var></var>')), [['<var>&lt;var onclick="alert()"&gt;</var>'], '</var>']);
+      assert.deepStrictEqual(inspect(parser('<ruby onclick="alert()">')), void 0);
+      assert.deepStrictEqual(inspect(parser('<ruby onclick="alert()"><ruby>')), void 0);
+      assert.deepStrictEqual(inspect(parser('<ruby><ruby onclick="alert()"></ruby></ruby>')), [['<ruby>&lt;ruby onclick="alert()"&gt;</ruby>'], '</ruby>']);
     });
 
     it('invalid', () => {
       assert.deepStrictEqual(inspect(parser('')), void 0);
-      assert.deepStrictEqual(inspect(parser('<var>')), void 0);
-      assert.deepStrictEqual(inspect(parser('<var>a')), void 0);
-      assert.deepStrictEqual(inspect(parser('</var>')), void 0);
+      assert.deepStrictEqual(inspect(parser('<ruby>')), void 0);
+      assert.deepStrictEqual(inspect(parser('<ruby>a')), void 0);
+      assert.deepStrictEqual(inspect(parser('</ruby>')), void 0);
       assert.deepStrictEqual(inspect(parser('a')), void 0);
-      assert.deepStrictEqual(inspect(parser('a<var>')), void 0);
+      assert.deepStrictEqual(inspect(parser('a<ruby>')), void 0);
     });
 
     it('ab', () => {
-      assert.deepStrictEqual(inspect(parser('<var></var>')), [['<var></var>'], '']);
-      assert.deepStrictEqual(inspect(parser('<var>a</var>')), [['<var>a</var>'], '']);
-      assert.deepStrictEqual(inspect(parser('<var>a</var>a')), [['<var>a</var>'], 'a']);
-      assert.deepStrictEqual(inspect(parser('<var>\n</var>')), [['<var></var>'], '']);
+      assert.deepStrictEqual(inspect(parser('<ruby></ruby>')), [['<ruby></ruby>'], '']);
+      assert.deepStrictEqual(inspect(parser('<ruby>a</ruby>')), [['<ruby>a</ruby>'], '']);
+      assert.deepStrictEqual(inspect(parser('<ruby>a</ruby>a')), [['<ruby>a</ruby>'], 'a']);
+      assert.deepStrictEqual(inspect(parser('<ruby>\n</ruby>')), [['<ruby></ruby>'], '']);
     });
 
     it('nest', () => {
-      assert.deepStrictEqual(inspect(parser('<var><var></var></var>')), [['<var><var></var></var>'], '']);
-      assert.deepStrictEqual(inspect(parser('<var>a<var>b</var>c</var>')), [['<var>a<var>b</var>c</var>'], '']);
-      assert.deepStrictEqual(inspect(parser('<var>`a`</var>')), [['<var><code>a</code></var>'], '']);
+      assert.deepStrictEqual(inspect(parser('<ruby><ruby></ruby></ruby>')), [['<ruby><ruby></ruby></ruby>'], '']);
+      assert.deepStrictEqual(inspect(parser('<ruby>a<ruby>b</ruby>c</ruby>')), [['<ruby>a<ruby>b</ruby>c</ruby>'], '']);
+      assert.deepStrictEqual(inspect(parser('<ruby>`a`</ruby>')), [['<ruby><code>a</code></ruby>'], '']);
     });
 
     it('escape', () => {
       assert.deepStrictEqual(inspect(parser('<a>')), void 0);
-      assert.deepStrictEqual(inspect(parser('<var><a></a></var>')), [['<var>&lt;a&gt;&lt;/a&gt;</var>'], '']);
-      assert.deepStrictEqual(inspect(parser('<var>a<a>b</a>c</var>')), [['<var>a&lt;a&gt;b&lt;/a&gt;c</var>'], '']);
+      assert.deepStrictEqual(inspect(parser('<ruby><a></a></ruby>')), [['<ruby>&lt;a&gt;&lt;/a&gt;</ruby>'], '']);
+      assert.deepStrictEqual(inspect(parser('<ruby>a<a>b</a>c</ruby>')), [['<ruby>a&lt;a&gt;b&lt;/a&gt;c</ruby>'], '']);
       assert.deepStrictEqual(inspect(parser('<img>')), void 0);
-      assert.deepStrictEqual(inspect(parser('<var><img></var>')), [['<var>&lt;img&gt;</var>'], '']);
+      assert.deepStrictEqual(inspect(parser('<ruby><img></ruby>')), [['<ruby>&lt;img&gt;</ruby>'], '']);
       assert.deepStrictEqual(inspect(parser('<img />')), void 0);
-      assert.deepStrictEqual(inspect(parser('<var><img /></var>')), [['<var>&lt;img /&gt;</var>'], '']);
+      assert.deepStrictEqual(inspect(parser('<ruby><img /></ruby>')), [['<ruby>&lt;img /&gt;</ruby>'], '']);
     });
 
     it('code', () => {